rename createPageContainer to newPageContainer
