+++ src/reducers.js
@@ -1,12 +1,12 @@
 import { combineReducers } from 'redux';
 import listsPage from './containers/ListsPageContainer/reducer';
 import listEditPage from './containers/ListEditPageContainer/reducer';
-import listCreatePage from './containers/ListCreatePageContainer/reducer';
+import listNewPage from './containers/ListNewPageContainer/reducer';
 
 const rootReducer = combineReducers({
   listsPage,
   listEditPage,
-  listCreatePage,
+  listNewPage,
 });
 
 export default rootReducer;
